AudioUtil: do not save a cover if it's already in cache Saving it again and again (in jpg) quicky decreases the cover quality

+++ vlc-android/src/org/videolan/vlc/gui/audio/AudioUtil.java
@@ -223,6 +223,8 @@ public class AudioUtil {
         OutputStream out = null;
         try {
             File file = new File(path);
+            if (file.exists() && file.length() > 0)
+                return;
             out = new BufferedOutputStream(new FileOutputStream(file), 4096);
             if (bitmap != null)
                 bitmap.compress(CompressFormat.JPEG, 90, out);
